% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_columns.R
\name{append_columns}
\alias{append_columns}
\title{Append columns so they retain tabulate attributes}
\usage{
append_columns(x, y)
}
\arguments{
\item{x}{tibble or data frame}

\item{y}{tibble or data frame}
}
\value{
tibble with tabulate attributes intact
}
\description{
Since tabulate utilises attributes assigned to columns, traditional methods of combining
may lose these attributes. This function will combine two structures together and reapply
the labels and tabulate attributes.
}
