% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_rows.R
\name{append_rows}
\alias{append_rows}
\title{Append rows so they retain tabulate attributes}
\usage{
append_rows(x, y)
}
\arguments{
\item{x}{tibble or data frame}

\item{y}{tibble or data frame}
}
\value{
tibble with tabulate attributes
}
\description{
Since tabulate utilises attributes assigned to columns, traditional methods of combining
will lose these attributes. This function will combine two structures together and reapply
the labels and tabulate attributes.
}
