% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach_attribute_vectors.R
\name{attach_attribute_vectors}
\alias{attach_attribute_vectors}
\title{Reattach tabulate attribute vectors to columns}
\usage{
attach_attribute_vectors(x, labels, tabulateParams)
}
\arguments{
\item{x}{tibble or data frame}

\item{labels}{labels to reattach}

\item{tabulateParams}{tabulate parameters to reattach}
}
\value{
tibble with tabulate attributes
}
\description{
Should we lose tabulate attributes (i.e. reading in a file) we can reattach the parameters
using this function.
}
