% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach_attributes.R
\name{attach_attributes}
\alias{attach_attributes}
\title{Reattach tabulate attributes to columns}
\usage{
attach_attributes(x, attributes)
}
\arguments{
\item{x}{tibble or data frame}

\item{attributes}{tibble with labels and tabulate columns}
}
\value{
tibble with tabulate attributes
}
\description{
Should we lose tabulate attributes (i.e. reading in a file) we can reattach the parameters
using this function.
}
