% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveytools.R
\name{barplot_by_column}
\alias{barplot_by_column}
\title{Barplotter by column}
\usage{
barplot_by_column(spread_sheet, file_name, vector_of_column_indexes)
}
\arguments{
\item{spread_sheet}{a data.frame
file_name to be used for the png file
vector_of_column_indexes a vector of column indexes (as in numbers)}
}
\value{
barplots named by column index, index_barplot.png
}
\description{
Takes a vector of column indexes, plots bars for each
}
\examples{
This could be an example
}
\keyword{barplot}
