% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveytools.R
\name{basicstatsmaker}
\alias{basicstatsmaker}
\title{Print basic stats}
\usage{
basicstatsmaker(spread_sheet, column_indexes)
}
\arguments{
\item{spread_sheet}{Our Qualtrics spreadsheet
column_indexes A vector of column names or numbers, columns must be factors}
}
\description{
Takes a spread_sheet, prints relevant stats for qualtrics survey
}
\keyword{not}
\keyword{why}
