% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{callback}
\alias{callback}
\title{Evaluate any JavaScript code after the \code{survey} object has been initialized}
\source{
\url{https://surveyjs.io/Documentation/Library?id=SurveyModel/}
}
\usage{
callback(id, JS)
}
\arguments{
\item{id}{widget or input$id}

\item{JS}{use \code{htmlwidgets::JS}}
}
\value{

}
\description{
Evaluate any JavaScript code after the \code{survey} object has been initialized
}
\examples{
# The `onValueChanged` api could be implemented via
\dontrun{

}
}
