% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{columns}
\alias{columns}
\title{Define columns for matrix questions}
\usage{
columns(id, columns)
}
\arguments{
\item{columns}{list of named list with names value and text}
}
\value{

}
\description{
Define columns for matrix questions
}
