% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_columns.R
\name{combine_columns}
\alias{combine_columns}
\title{Combine columns}
\usage{
combine_columns(.data, new_column_name, pattern, type = "starts_with")
}
\arguments{
\item{.data}{Dataframe to combine column in}

\item{new_column_name}{New column name}

\item{pattern}{pattern to match against}

\item{type}{one of "starts_with", "ends_with", "matches", "contains", "one_of". Defaults to starts_with}
}
\value{
TBU
}
\description{
TBU
}
