% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveytools.R
\name{csvcleaner}
\alias{csvcleaner}
\title{CSV cleaner}
\usage{
csvcleaner(spread_sheet)
}
\arguments{
\item{spread_sheet}{a R data.frame}
}
\value{
a clean data.frame
}
\description{
Takes out all the stuff from MechTurk and Qualtrics we don't want
}
\examples{
This could be an example
}
\keyword{cats}
