% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decrypt_data.R
\name{decrypt_data}
\alias{decrypt_data}
\title{Decrypt an encrypted .RData file}
\usage{
decrypt_data(fileLocation)
}
\arguments{
\item{fileLocation}{location of the file}
}
\description{
A shiny gadget will pop up so the password can be provided
}
