% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.r
\docType{data}
\name{.log_env}
\alias{.log_env}
\title{Logging}
\format{
An object of class \code{environment} of length 2.
}
\usage{
.log_env
}
\description{
Initialising a new logging env to save outputs in
}
\keyword{datasets}
