% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encrypt_data.R
\name{encrypt_data}
\alias{encrypt_data}
\title{Encrypt and save and .RData file}
\usage{
encrypt_data(data, password, saveLocation)
}
\arguments{
\item{data}{data structure to encrypt}

\item{password}{password to encrypt/decrypt the file}

\item{saveLocation}{location to save the .RData file}
}
\description{
Encrypt and save and .RData file
}
