% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_segments.R
\name{extract_segments}
\alias{extract_segments}
\title{Extract segment columns from tabulate data structure}
\usage{
extract_segments(x)
}
\arguments{
\item{x}{tabulate data structure with tabulate parameters}
}
\value{
tibble with columns labelled as "segment" in tabulate parameters
}
\description{
Extract segment columns from tabulate data structure
}
