% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/fetch_var_by.R
\name{fetch_var_by}
\alias{fetch_var_by}
\title{Fetches the values of one variable where a second variable is a certain value.}
\usage{
fetch_var_by(first_var, by_var, select, data)
}
\arguments{
\item{first_var}{character. The name of the variable you want values returned for.}

\item{by_var}{character. The name of the variable you want to sort by.}

\item{select.}{The value by_var should be.}

\item{data.}{The data to grap values from. This is assumed to be a three column
melted dataframe, using reshape2 <http://cran.r-project.org/web/packages/reshape2/index.html>.
The first column should contain IDs, the second column should contain variable
names, and the third column should contain variable values.

e.g., to return the favorite food of people from the city of Paris:
\code{fetch_var_by(favorite_food, home_city, 'Paris', data)}}
}
\description{
Fetches the values of one variable where a second variable is a certain value.
}

