% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_attributes.R
\name{get_attributes}
\alias{get_attributes}
\title{Obtain tabulate attributes}
\usage{
get_attributes(x, export = FALSE,
  filename = sprintf("Survey Tools Parameters \%s.csv", Sys.Date()))
}
\arguments{
\item{x}{list}

\item{export}{logical indicating whether we want to export the parameters into a csv file within the current working directory}

\item{filename}{string defaulting to "Survey Tools Parameters Sys.Date().csv"}
}
\value{
tibble of parameters
}
\description{
Should we want to view the tabulate attributes as a tibble, in case of editing or
checking accuracy, we can use this function to extract them and store within a tibble.
}
\examples{
library(surveySimulate)
example <- simulateSurvey(100)
parameters <- get_attributes(example)
}
