% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{grouper}
\alias{grouper}
\title{Grouper}
\usage{
grouper(.data)
}
\arguments{
\item{.data}{Dataframe to group and count by}
}
\value{
Dataframe of columns and counts
}
\description{
Wrapper around count functions
}
