% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_attributes.R
\name{import_attributes}
\alias{import_attributes}
\title{Import tabulate attributes from a csv file}
\usage{
import_attributes(data, fileLocation)
}
\arguments{
\item{data}{data structure to apply the parameters to}

\item{fileLocation}{string indicating location of tabulate attributes file}
}
\value{
tibble containing tabulate attributes
}
\description{
If we have exported tabulate attributes for editing in a .csv file then we can use this function to import them and also append them to a specified data structure.
}
