% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_table.R
\name{index_table}
\alias{index_table}
\title{Construct indexed tables from a tabulate generated table}
\usage{
index_table(x, baseSegment, order = TRUE)
}
\arguments{
\item{x}{Table generated from tabulate}

\item{baseSegment}{Segment name to treat as base for indexing}

\item{order}{Do we want to order the indexed data by segment in order of index extremity}
}
\value{
tidy tibble with indexed and classified responses
}
\description{
DESCRIPTION
}
