% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.r
\name{log_init}
\alias{log_init}
\title{Initialise logging}
\usage{
log_init(path)
}
\arguments{
\item{path}{path / filename of the log to initialise}
}
\description{
Initialise html log by provide path to log. No extension is needed - will automate to html.
}
