% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.r
\name{log_print}
\alias{log_print}
\title{Log Print}
\usage{
log_print(log = NULL)
}
\arguments{
\item{log}{Name of log to force print. Will default to last value}
}
\description{
Force render rmarkdown to print html
}
