% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.r
\name{log_render}
\alias{log_render}
\title{Render log}
\usage{
log_render(log_name = "default")
}
\arguments{
\item{log_name}{Name of the log to render. Defaults to last called log file}
}
\description{
Render the definied log file using rmarkdown with default options
}
