% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.r
\name{log_table}
\alias{log_table}
\alias{log_table.default}
\alias{log_table.svyt_df}
\title{Log table using formattable}
\usage{
log_table(
  .data,
  header = "",
  ...,
  log = NULL,
  print = getOption("surveytools.force_print")
)

\method{log_table}{default}(
  .data,
  header = "",
  ...,
  log = NULL,
  print = getOption("surveytools.force_print")
)

\method{log_table}{svyt_df}(
  .data,
  header = "",
  ...,
  log = NULL,
  print = getOption("surveytools.force_print")
)
}
\arguments{
\item{.data}{Dataframe or summarised outputs to log to table}

\item{header}{Header for the table}

\item{...}{additional options passed to formattable::format_table()}

\item{log}{Overwrite default log that gets logged to and rendered}
}
\description{
Log a table of results using formattable. Has custom methods for the common summariser outputs as well as a generic method for all tables
}
