% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_rdsDMU.R
\name{make_rdsDMU}
\alias{make_rdsDMU}
\title{Function for generating standardised .rds DMU formats}
\usage{
make_rdsDMU(DMUName, cat, summary, usrGen, contactGen, scriptGen, version,
  coldescription, colucd, colunits, README, test = FALSE)
}
\arguments{
\item{DMUName}{Name for the DMU}

\item{cat}{data.frame of catalogue data to add to DMU}

\item{summary}{a 1/2 sentence overview description of your DMU}

\item{usrGen}{name of person who is making DMU}

\item{contactGen}{email contact of person who is making DMU}

\item{scriptGen}{script used to generate DMU}

\item{version}{version number identifier of script}

\item{coldescription}{a vector containing string descriptions of each column in cat}

\item{colucd}{a vector containing string Universal Content Descriptors (UCDs) for each column in cat (see http://cdsweb.u-strasbg.fr/UCD/tree/js/)}

\item{colunits}{a vector containing string units for each column in cat}

\item{README}{# a string README for the catalogue (please separate lines with \n}

\item{test}{TRUE/FLASE - If true catalogue can be run with dummy values (see example) and not give errors}
}
\value{
a list containing the DMU. The DMU will also be written out as a .rds file with the format: DMUName,'_',format(Sys.time(), "%d_%m_%Y"),'_v',version,'.rds'
}
\description{
Takes a catalogue and various meta data/column information and writes out a standardised Data Management Unit (DMU) data product as a .rds file.
}
\author{
L. Davies <luke.j.davies@uwa.edu>
}
