% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_element.R
\name{name_element}
\alias{name_element}
\title{Place name of list as attribute within the element}
\usage{
name_element(x)
}
\arguments{
\item{x}{list}
}
\value{
list with names as attributes
}
\description{
Where we want the name of the list to be accessible within a loop or function, we can
use this function to place the name within the list element as an attribute.
}
