% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option_breaker.R
\name{option_breaker}
\alias{option_breaker}
\title{Option breaker}
\usage{
option_breaker(.data, str, to_check, group = "")
}
\arguments{
\item{.data}{Dataframe}

\item{str}{TBU}

\item{to_check}{TBU}

\item{group}{TBU}
}
\value{
TBU
}
\description{
TBU
}
