% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveytools.R
\name{ordermachine}
\alias{ordermachine}
\title{Order machine to get our labels right}
\usage{
ordermachine(spread_sheet)
}
\arguments{
\item{Our}{Qualtrics spreadsheet}
}
\description{
Takes a spread_sheet, orders the factors as indicated.
}
\keyword{things}
\keyword{ugly}
