% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/percent_table.R
\name{percent_table}
\alias{percent_table}
\title{Prints a percent table.}
\usage{
percent_table(df, round = NULL)
}
\arguments{
\item{df}{data frame. The data frame of the variable to run the table for.  This dataframe is still assumed
to be three column (id, variable name, value), but could be the result of a fetch_var() or fetch_var_by()}

\item{round}{integer. How many digits you want to round percentages to. Default no rounding.}

\item{var}{characer. The name of the variable you want percent breakdown for.}
}
\description{
Prints a percent table.
}

