% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.ssize.R
\name{print.ssize}
\alias{print.ssize}
\title{print.ssize}
\usage{
\method{print}{ssize}(x)
}
\arguments{
\item{x}{Class 'ssize' object to print}
}
\value{

}
\description{
Print objects of class 'ssize' generated by \code{\link{survey_size}}.
}
\examples{
\dontrun{
library(surveytools)
foo <- survey_size(population = 400)
print(foo)
}
}

