% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdsToFitsDMU.R
\name{rdsToFitsDMU}
\alias{rdsToFitsDMU}
\title{Function for convert DMUs generated by make_rdsDMU() into FITS format}
\usage{
rdsToFitsDMU(DMU)
}
\arguments{
\item{DMU}{Object generated by make_rdsDMU()}
}
\value{
a FITS file with the format: DMUName,'_',format(Sys.time(), "%d_%m_%Y"),'_v',version,'.fits'
}
\description{
Takes a catalogue generated using make_rdsDMU() and make it into a sensible FITS binary table.
}
\author{
L. Davies <luke.j.davies@uwa.edu>
}
