% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sav_labelled.R
\name{read_sav_labelled}
\alias{read_sav_labelled}
\title{Read in .sav file with labelled columns}
\usage{
read_sav_labelled(file)
}
\arguments{
\item{file}{file location}
}
\value{
tibble
}
\description{
Read in .sav file with labelled columns
}
