% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveytools.R
\name{relativeplotter}
\alias{relativeplotter}
\title{Plot relative to control stimulus}
\usage{
relativeplotter(control_stimulus, data_set, vector_of_columns, filename)
}
\arguments{
\item{control_stimulus}{The control stimulus to create y=0
data_set The data to plot relative to the control_stimulus
vector_of_columns A vector of column names to be plotted
filename A string determining the name of the output .png file}
}
\description{
Plots treatment stimuli in relation to control stimulus.
}
\keyword{not}
\keyword{why}
