% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{rows}
\alias{rows}
\title{Define rows for matrix questions}
\usage{
rows(id, rows)
}
\arguments{
\item{rows}{list of named list with names value and text}
}
\value{

}
\description{
Define rows for matrix questions
}
