% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveytools.R
\name{spreadsheetbuilder}
\alias{spreadsheetbuilder}
\title{Build your own spreadsheet}
\usage{
spreadsheetbuilder(vector_of_files)
}
\arguments{
\item{vector_of_files}{a vector of file names to be read with read.csv()}
}
\value{
spread sheet containing all the data contained in vector_of_files
}
\description{
Takes a vector of files and combines their content into a big spreadsheet.
Uses read.csv to read, csvcleaner to clean, rbind to combine. 
Add columns 'stimulus' that takes its value from the filename.
}
\examples{
This could be an example
}
\keyword{builder}
