% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariser_base.R
\name{summariser_base}
\alias{summariser_base}
\alias{summariser_base.default}
\title{Summarise survey data}
\usage{
summariser_base(
  .data,
  var,
  ...,
  stat = c("mean", "median", "total"),
  compare = "overall",
  survey.design = NULL,
  ci_level = 0.95,
  simplify = TRUE
)

\method{summariser_base}{default}(
  .data,
  var,
  ...,
  stat = c("mean", "median", "total"),
  compare = "overall",
  survey.design = NULL,
  ci_level = 0.95,
  simplify = TRUE
)
}
\arguments{
\item{.data}{Dataframe or survey data to be summarised}

\item{var}{Variable within dataframe to be summarised}

\item{...}{Variables to group by}

\item{stat}{One of "mean", "median" or "total". Defaults to mean if not correctly provided}

\item{compare}{Stat to use for significance tests of survey results. Must be provided in "group::response" format. Defaults to the overall value}

\item{survey.design}{List of survey design inputs to be applied to dataframe, if not survey data of class 'tbl_svy'. Please see '?survey::svydesign' to identify required inputs}

\item{ci_level}{confidence level to be summarised for}

\item{simplify}{Should the survey results be simplified from decimals to numbers?}
}
\value{
A dataframe of summarised results
}
\description{
This function provides a summary of survey data based on the design provided,
as well as checks for significance of results
}
