% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariser.R
\name{summariser_character}
\alias{summariser_character}
\title{summariser_character}
\usage{
summariser_character(.data, var, group, stat, level)
}
\arguments{
\item{.data}{Dataframe or survey data to be summarised}

\item{var}{Variable within dataframe to be summarised}

\item{group}{Variables to group by}

\item{stat}{One of "mean", "median" or "total". Defaults to mean if not correctly provided}

\item{level}{confidence level to be summarised for}
}
\value{
A dataframe of summarised results for numeric variables

#' @importFrom magrittr "\%>\%"
}
\description{
Internal function to summarise character variables
}
