% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate_columns.R
\name{tabulate_columns}
\alias{tabulate_columns}
\title{Return table of responses to ordinal/nominal response questions}
\usage{
tabulate_columns(x, weights = NULL, segments = NULL, filterNA = TRUE,
  binary = FALSE, multipleChoice = FALSE)
}
\arguments{
\item{x}{tibble of ordinal/nominal responses}

\item{weights}{(Optional) vector of weights}

\item{segments}{(Optional) vector or data frame of segment membership}

\item{filterNA}{Whether we remove NAs before calculating percentage}

\item{binary}{Are the columns binary responses (1 or TRUE)}

\item{multipleChoice}{Are the columns multiple choice and therefore should be grouped together.}
}
\value{
tidy tibble containing percentage table of responses
}
\description{
This function returns a tidy tibble containing percentage tables
}
