% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate_data.R
\name{tabulate_data}
\alias{tabulate_data}
\title{Construct tables from a tabulate data structure.}
\usage{
tabulate_data(x, weights = NULL, segments = NULL, filterNA = TRUE,
  addBase = FALSE)
}
\arguments{
\item{x}{tabulate data structure with labels and tabulate parameters}

\item{weights}{weights column}

\item{segments}{segments column(s)}

\item{filterNA}{boolean indcating whether NA's are filtered prior to percentage calculation}

\item{addBase}{boolean indicating whether a column "Base" should be added}
}
\value{
Tidy tibble containing the response to questions
}
\description{
Taking a data structure with labels and parameters this function will return a tidy tibble that is a breakdown of responses to each question within the structure. These tables can be weighted or segmented by providing the columns and can also contain a Base column which is useful for indexing.
}
