% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/significance_tests.R
\name{test_significance}
\alias{test_significance}
\title{test_significance}
\usage{
test_significance(.data, compare = compare)
}
\arguments{
\item{.data}{Summarised dataframe}

\item{compare}{Group::response combination to be used for comparison}
}
\value{
A dataframe of summarised results for numeric variables

#' @importFrom magrittr "\%>\%"
}
\description{
Internal function to test significance of results
}
