% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/ttest.R
\name{ttest}
\alias{ttest}
\title{Runs a t test (two-sided, unpaired) on data within a melted dataframe.}
\usage{
ttest(var, by_var, ignore = NULL, data = get_data())
}
\arguments{
\item{var}{character. The name of the first variable in the chi-square test.}

\item{by_var}{charager. The name of the second (by) variable in the
chi-square test.}

\item{ignore}{vector. A vector of the names of values to be replaced by NAs
    within both the var and the by_var. Can clean up categories that are too
    small for meaningful analysis.}

\item{data.}{The data to be transformed.  This is assumed to be a three
column melted dataframe, using reshape2 <http://cran.r-project.org/web/packages/reshape2/index.html>.
The first column should contain IDs, the second column should contain variable
names, and the third column should contain variable values.}
}
\description{
Gracefully handles differences in variable size and NAs, and allows easy
removal of variables.
}

