% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/significance_tests.R
\name{welch_ttest}
\alias{welch_ttest}
\title{Welch T test for unequal variances}
\usage{
welch_ttest(m1, m2, s1, s2, n1, n2, m0 = 0, equal.variance = FALSE)
}
\arguments{
\item{m1}{Mean of group 1}

\item{m2}{Mean of group 2}

\item{s1}{Standard deviation of group 1}

\item{s2}{Standard deviation of group 2}

\item{n1}{N for group 1}

\item{n2}{N for group 2}

\item{m0}{Diff of means to be compared against. Defaults to 0}

\item{equal.variance}{Are variances of the two groups to be taken as the same}
}
\value{
A dataframe of pvalues and means
}
\description{
Internal function to summarise numeric variables
}
