% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.r
\name{log_init}
\alias{log_init}
\title{Function to initialise a new logger - defines the output file and options available, if any}
\usage{
log_init(path)
}
\description{
Function to initialise a new logger - defines the output file and options available, if any
}
